Using my fork of metalsmith that adds list of files to exclude from clean
This allows us to regenerate the static website everytime, removing the files that have to be removed, like pages that don't have a source anymore, while keeping all the important files in the root folder.

This is all because the GitHub pages read from the root folder of the repo.